+++ rich/spinner.py
@@ -1,25 +1,24 @@
 from typing import cast, List, Optional, TYPE_CHECKING
 
 from ._spinners import SPINNERS
-from .console import Console
 from .measure import Measurement
-from .style import StyleType
 from .text import Text, TextType
 
 if TYPE_CHECKING:
     from .console import Console, ConsoleOptions, RenderResult
+    from .style import StyleType
 
 
 class Spinner:
     def __init__(
-        self, name: str, text: TextType = "", *, style: StyleType = None, speed=1.0
+        self, name: str, text: TextType = "", *, style: "StyleType" = None, speed=1.0
     ) -> None:
         """A spinner animation.
 
         Args:
             name (str): Name of spinner (run python -m rich.spinner).
             text (TextType, optional): Text to display at the right of the spinner. Defaults to "".
-            style (StyleType, optional): Style for sinner amimation. Defaults to None.
+            style (StyleType, optional): Style for spinner animation. Defaults to None.
             speed (float, optional): Speed factor for animation. Defaults to 1.0.
 
         Raises:
@@ -35,7 +34,7 @@ class Spinner:
         self.start_time: Optional[float] = None
         self.style = style
         self.speed = speed
-        self.time = 0.0
+        self._update_speed = 0.0
 
     def __rich_console__(
         self, console: "Console", options: "ConsoleOptions"
@@ -62,8 +61,29 @@ class Spinner:
             Text: A Text instance containing animation frame.
         """
         frame_no = int((time * self.speed) / (self.interval / 1000.0))
+        if self._update_speed:
+            self.start_time += time - (time * self.speed / self._update_speed)
+            self.speed = self._update_speed
+            self._update_speed = 0.0
         frame = Text(self.frames[frame_no % len(self.frames)], style=self.style or "")
         return Text.assemble(frame, " ", self.text) if self.text else frame
+
+    def update(
+        self, *, text: TextType = "", style: "StyleType" = None, speed: float = None
+    ) -> None:
+        """Updates attributes of a spinner after it has been started.
+
+        Args:
+            text (TextType, optional): Text to display at the right of the spinner. Defaults to "".
+            style (StyleType, optional): Style for spinner animation. Defaults to None.
+            speed (float, optional): Speed factor for animation. Defaults to None.
+        """
+        if text:
+            self.text = text
+        if style:
+            self.style = style
+        if speed:
+            self._update_speed = speed
 
 
 if __name__ == "__main__":  # pragma: no cover