+++ tests/test_spinner.py
@@ -1,4 +1,3 @@
-from time import time
 import pytest
 
 from rich.console import Console
@@ -7,8 +6,7 @@ from rich.spinner import Spinner
 
 
 def test_spinner_create():
-    spinner = Spinner("dots")
-    assert spinner.time == 0.0
+    Spinner("dots")
     with pytest.raises(KeyError):
         Spinner("foobar")
 
@@ -31,6 +29,28 @@ def test_spinner_render():
     result = console.end_capture()
     print(repr(result))
     expected = "⠋ Foo\n⠙ Foo\n"
+    assert result == expected
+
+
+def test_spinner_update():
+    time = 0.0
+
+    def get_time():
+        nonlocal time
+        return time
+
+    console = Console(width=80, force_terminal=True, get_time=get_time)
+    console.begin_capture()
+    spinner = Spinner("dots", "Foo")
+    console.print(spinner)
+    spinner.update(text="Bar", style="green", speed=2)
+    time += 80 / 1000
+    console.print(spinner)
+    time += 80 / 1000
+    console.print(spinner)
+    result = console.end_capture()
+    print(repr(result))
+    expected = f"⠋ Foo\n\x1b[32m⠙\x1b[0m Bar\n\x1b[32m⠸\x1b[0m Bar\n"
     assert result == expected